@@ -42,7 +42,7 @@ class TestMultiForm(AppDataTestCase):
             'fs-0-myapp-title': 'First',
             'fs-0-myapp-publish_from': '2010-11-12',
         }
-        formset = FormSet(data, prefix='fs')
+        formset = FormSet('Article', data, prefix='fs')
 
         tools.assert_true(formset.is_valid())
         formset.save()
@@ -67,7 +67,7 @@ class TestMultiForm(AppDataTestCase):
             'myapp-publish_from': '2010-11-12',
             'myapp2-foo': 'Second',
         }
-        form = MF(data)
+        form = MF('Article', (), data)  # Does this do what it should though? It stops the error, but...??
         tools.assert_true(form.is_valid())
         tools.assert_equals({}, form.errors)
         art = form.save()
@@ -91,7 +91,7 @@ class TestMultiForm(AppDataTestCase):
             'myapp-title': 'First',
             'myapp-publish_from': '2010-11-12',
         }
-        form = MF(data)
+        form = MF('Article', (), data)  # Does this do what it should though? It stops the error, but...??
         tools.assert_true(form.is_valid())
         tools.assert_equals({}, form.errors)
         art = form.save()
@@ -114,7 +114,7 @@ class TestMultiForm(AppDataTestCase):
             'myapp-title': 'First',
             'myapp-publish_from': '2010-11-12',
         }
-        form = MF(data)
+        form = MF('Article', (), data)  # Does this do what it should though? It stops the error, but...??
         tools.assert_true(form.is_valid())
         tools.assert_equals({}, form.errors)
         art = form.save()
@@ -145,7 +145,7 @@ class TestMultiForm(AppDataTestCase):
             'myapp-title': 'First',
             'myapp-publish_from': '2010-11-12',
         }
-        form = MF(data)
+        form = MF('Article', (), data)  # Does this do what it should though? It stops the error, but...??
         tools.assert_true(form.is_valid())
         tools.assert_equals({}, form.errors)
         art = form.save()
